# backend/docs/swagger.yaml
definitions:
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ApparatResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      short_name:
        type: string
      system_parts:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SystemPartResponse'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectInput:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      description:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        maximum: 255
        minimum: 0
        type: integer
      hardware_type:
        type: string
      notification_class_id:
        type: string
      optional:
        type: boolean
      software_number:
        maximum: 65535
        minimum: 0
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        maxLength: 250
        type: string
      text_individual:
        maxLength: 250
        type: string
    required:
    - software_number
    - software_type
    - text_fix
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      field_device_id:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        type: integer
      hardware_type:
        type: string
      id:
        type: string
      notification_class_id:
        type: string
      optional:
        type: boolean
      software_number:
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        type: string
      text_individual:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateObjectDataRequest:
    properties:
      apparat_ids:
        items:
          type: string
        type: array
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectInput'
        type: array
      description:
        maxLength: 250
        type: string
      is_active:
        type: boolean
      project_id:
        type: string
      version:
        maxLength: 100
        type: string
    required:
    - description
    - version
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ObjectDataResponse:
    properties:
      apparats:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ApparatResponse'
        type: array
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      project_id:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SystemPartResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      short_name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateObjectDataRequest:
    properties:
      apparat_ids:
        items:
          type: string
        type: array
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectInput'
        type: array
      description:
        type: string
      is_active:
        type: boolean
      project_id:
        type: string
      version:
        type: string
    type: object
...
  internal_handler_facility.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  internal_handler_facility.ObjectDataListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ObjectDataResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handler_facility.ObjectDataResponse:
    properties:
      apparats:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ApparatResponse'
        type: array
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      project_id:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/facility/object-data:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter by Apparat ID
        in: query
        name: apparat_id
        type: string
      - description: Filter by System Part ID
        in: query
        name: system_part_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_facility.ObjectDataListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler_facility.ErrorResponse'
      summary: List object data with pagination
      tags:
      - facility-object-data
    post:
      consumes:
      - application/json
      parameters:
      - description: Object Data
        in: body
        name: object_data
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateObjectDataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ObjectDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Create object data
      tags:
      - facility-object-data
  /api/v1/facility/object-data/{id}:
    delete:
      parameters:
      - description: Object Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Delete object data
      tags:
      - facility-object-data
    get:
      parameters:
      - description: Object Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_facility.ObjectDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler_facility.ErrorResponse'
      summary: Get object data by ID
      tags:
      - facility-object-data
    put:
      consumes:
      - application/json
      parameters:
      - description: Object Data ID
        in: path
        name: id
        required: true
        type: string
      - description: Object Data
        in: body
        name: object_data
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateObjectDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ObjectDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Update object data
      tags:
      - facility-object-data
  /api/v1/facility/object-data/{id}/bacnet-objects:
    get:
      parameters:
      - description: Object Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Get bacnet objects for object data
      tags:
      - facility-object-data
...
swagger: "2.0"